backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pdfplumber==0.11.5
groq==0.15.0
reportlab==4.2.5
python-multipart==0.0.20
httpx==0.28.1
python-dotenv==1.0.1
pytest==8.3.5
